refactor: delete addtional features(congestion-graph, lostnfound)
