+++ frontend/app/src/main/java/com/example/shattle/MainActivity.kt
@@ -47,10 +47,8 @@ class MainActivity : AppCompatActivity() {
         binding.infoImageButton.setOnClickListener {
             // 다른 팝업이 열려있는지 확인 (닫혀있다면 null 이 할당됨)
             val existingFragmentInfo = supportFragmentManager.findFragmentByTag("InfoFragment")
-            val existingFragmentChart =
-                supportFragmentManager.findFragmentByTag("CongestionGraphFragment")
             // 다른 팝업이 닫혀 있는 경우에만 열기
-            if (existingFragmentInfo == null && existingFragmentChart == null) {
+            if (existingFragmentInfo == null) {
                 val infoFragment = InfoFragment()
                 val transaction = supportFragmentManager.beginTransaction()
                 // infoFragment 가 표시될 위치 설정 (infoContainer), tag 설정(tag 로 fragment 가 열려있는지 확인 가능)
