+++ frontend/app/src/main/java/com/example/shattle/ui/timetable/TimetableViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.shattle.ui.lostnfound
+package com.example.shattle.ui.timetable
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
