--- frontend/app/src/main/res/layout/fragment_congestion_graph.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.congestion.CongestionGraphFragment"
-    android:layout_marginTop="@dimen/top_container_height">
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/popUpContainer"
-        android:layout_width="match_parent"
-        android:layout_height="400dp"
-        android:layout_margin="30dp"
-        app:cardCornerRadius="20dp"
-        app:cardElevation="16dp"
-        app:contentPadding="5dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ImageButton
-            android:id="@+id/closeImageButton"
-            android:layout_width="@dimen/button_size"
-            android:layout_height="@dimen/button_size"
-            android:background="@color/transparent"
-            android:src="@drawable/btn_arrow_back"
-            app:tint="@color/gray" />
-
-        <com.github.mikephil.charting.charts.LineChart
-            android:id="@+id/dayTimeChart"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:visibility="gone" />
-    </androidx.cardview.widget.CardView>
-</androidx.constraintlayout.widget.ConstraintLayout>
