--- frontend/app/src/main/res/layout/fragment_lost_item.xml
@@ -1,34 +1,0 @@
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cardView"
-    android:layout_width="match_parent"
-    android:layout_height="170dp"
-    app:cardCornerRadius="10dp"
-    app:cardElevation="@dimen/card_view_elevation"
-    app:cardPreventCornerOverlap="false"
-    app:cardUseCompatPadding="true"
-    android:layout_marginVertical="20dp"
-    android:layout_marginHorizontal="20dp">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal"
-        android:padding="20dp">
-
-        <TextView
-            android:id="@+id/itemNameTextView"
-            android:layout_width="100dp"
-            android:layout_height="match_parent"
-            android:text="Item_Title"
-            android:textSize="18sp" />
-
-        <TextView
-            android:id="@+id/itemDescriptionTextView"
-            android:layout_width="200dp"
-            android:layout_height="match_parent"
-            android:text="Item_Description" />
-
-    </LinearLayout>
-</androidx.cardview.widget.CardView>
