+++ frontend/app/src/main/res/layout/fragment_station.xml
@@ -141,26 +141,4 @@
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <ImageButton
-        android:id="@+id/chartButton"
-        android:layout_width="@dimen/button_size"
-        android:layout_height="@dimen/button_size"
-        android:background="@color/transparent"
-        android:scaleType="fitXY"
-        android:src="@drawable/btn_chart_2"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.919"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/waitingLineTextView"
-        app:layout_constraintVertical_bias="0.849" />
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/congestionGraphContainer"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
